requirements.txt
# type hint syntax with | introduced in python 3.10
python>=3.10
# np.default_rng() introduced in numpy 1.17
numpy>=1.17
pandas
geopandas>=0.14.4
scikit-learn
matplotlib
ipykernel
contextily
pyarrow
python-dotenv